lint

+++ clickhouse_ping.go
@@ -1,13 +1,13 @@
 package clickhouse
 
 func (ch *clickhouse) ping() error {
-	ch.log("-> ping")
+	ch.logf("-> ping")
 	if err := writeUvarint(ch.conn, ClientPingPacket); err != nil {
 		return err
 	}
 	if err := ch.gotPacket(ServerPongPacket); err != nil {
 		return err
 	}
-	ch.log("<- pong")
+	ch.logf("<- pong")
 	return nil
 }
